requirements.txt
pandas
feast[redis]
fastapi
uvicorn
python-multipart
aiofiles
great-expectations
prometheus-client
pydantic
bentoml
tqdm
mlflow
numpy
lightfm
scikit-learn
jupyter
matplotlib
kafka-python
